+++ .gitignore
@@ -3,3 +3,5 @@
 .gradle/
 build/
 classes/
+project/
+target/
