--- src/main/java/uk.gov.rifs/Application.java
@@ -1,26 +1,0 @@
-package uk.gov.rifs.application.controller;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
-
-@SpringBootApplication
-public class Application {
-
-    private static final Logger log = LoggerFactory.getLogger(Application.class);
-
-    public static void main(String[] args) {
-        System.out.println("Spring Boot Application Started ............");
-        SpringApplication.run(Application.class, args);
-    }
-
-
-}
-
-
