+++ src/main/java/uk/gov/rifs/Application.java
@@ -1,0 +1,18 @@
+package uk.gov.rifs;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+    private static final Logger log = LoggerFactory.getLogger(Application.class);
+
+    public static void main(String[] args) {
+        System.out.println("Spring Boot Application Started ............");
+        SpringApplication.run(Application.class, args);
+    }
+}
+
+
